requirements.txt
clickclick>=1.0
boto3>=1.2.3
botocore>=1.3.22
netaddr>=0.7.18
requests>=2.9.1
aws-saml-login>=1.0.7
keyring
python-gnupg>=0.3.7
python-dateutil
datetime
PyYAML
